requirements.txt
Pillow>=11.1.0
numpy>=2.2.4 